Optimization

To demonstrate the package's capabilities, let's apply it to the Rosenbrock function, a well-known benchmark in numerical optimization

Defining a Problem

To optimize a function using HierarchicMemeticStrategy.jl, you must encapsulate your objective function and its constraints within a FunctionProblem object.

First define the Rosenbrock function:

julia> using HierarchicMemeticStrategy
julia> rosenbrock(x) = (1.0 - x[1])^2 + 100.0 * (x[2] - x[1]^2)^2;

Next, wrap it in a FunctionProblem object, specifying the bounds for each variable and whether to maximize or minimize:

julia> lower_bounds = [-5.0, -5.0];
julia> upper_bounds = [5.0, 5.0];
julia> problem = FunctionProblem(
           fitness_function = rosenbrock,
           lower = lower_bounds,
           upper = upper_bounds,
           maximize = false
       );

HierarchicMemeticStrategy.FunctionProblem — Type

FunctionProblem(; fitness_function, lower, upper, maximize=false)

Construct an optimization problem for a standard mathematical function.

Arguments

fitness_function::Function: The objective function to evaluate. It must accept a single vector x and return a scalar numerical value.

lower::Vector: A vector specifying the minimum allowable values for each dimension.

upper::Vector: A vector specifying the maximum allowable values for each dimension.

maximize::Bool: (Optional) Set to true for maximization, or false for minimization (default).

Examples

problem = FunctionProblem(
    fitness_function = x -> x[1]^2 + x[2]^2,
    lower = [-5.0, -5.0],
    upper = [5.0, 5.0]
)

Performing Optimization

Once Function Problem is defined, we can perform optimization using hms algorithm.

julia> res = hms(optimization_problem=problem)┌────────────────────── HMSResult Summary ───────────────────────
│ • Metaepoch Count:      16
│ • Solution:             [0.9999999926684017, 0.9999999853317146]
│ • Best Fitness:         5.37549235437392e-17
│ • Fitness Evaluations:  11924
└────────────────────────────────────────────────────────────────

Configuration

The example above uses default configuration. For more configuration details check HMS Configuration section.

Obtaining Results

Once the optimization process is complete, the solver returns an HMSResult object. This object contains the full history of the run, the best found solution, and tools for visualization.

List of functions

HMSResult interface for optimization result.

HierarchicMemeticStrategy.HMSResult — Type

HMSResult

A structure containing the results of an HMS optimization run.

HierarchicMemeticStrategy.summary — Method

summary(result::HMSResult)

Print a formatted summary table to the standard output for a quick overview of the results.

HierarchicMemeticStrategy.solution — Method

solution(result::HMSResult)

Return the best-found input vector (solution).

HierarchicMemeticStrategy.best_fitness — Method

best_fitness(result::HMSResult)

Returns the scalar fitness value at the best-found point.

HierarchicMemeticStrategy.iterations — Method

iterations(result::HMSResult)

Return the total count of metaepochs performed during the run.

HierarchicMemeticStrategy.f_calls — Method

f_calls(result::HMSResult)

Return the total number of times the objective function was evaluated across all demes.

HierarchicMemeticStrategy.metaepoch_data — Method

metaepoch_data(result::HMSResult)

Return algorithm data.

HierarchicMemeticStrategy.plotPopulations — Method

plotPopulations(result::HMSResult)

Show interactive plot presenting changes in populations during metaepochs.

Arguments

result::HMSResult: The result object returned by the optimizer.

x_index::Int: The index of the dimension to plot on the x-axis (default: 1).

y_index::Int: The index of the dimension to plot on the y-axis (default: 2).

HierarchicMemeticStrategy.plotBestFitness — Method

plotBestFitness(result::HMSResult)

Plot best_fitness value during metaepochs.

HierarchicMemeticStrategy.plotDeme — Method

 plotDeme(result::HMSResult)

Plot single Deme population over metaepochs.

Arguments

result::HMSResult: The result object returned by the HMS optimizer.

deme_index::Int: The index of deme to visualize (default: 1).

x_index::Int: The index of the dimension to plot on the x-axis (default: 1).

y_index::Int: The index of the dimension to plot on the y-axis (default: 2).